Criação de exceções personalizadas; Tratamento de exceções com @ControllerAdvice e @Exception Handler

+++ src/main/java/LuanFigueiraOliveira/ECommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package LuanFigueiraOliveira.ECommerce.services;
 import LuanFigueiraOliveira.ECommerce.dto.ProductDTO;
 import LuanFigueiraOliveira.ECommerce.entities.Product;
 import LuanFigueiraOliveira.ECommerce.repositories.ProductRepository;
+import LuanFigueiraOliveira.ECommerce.services.exceptions.DatabaseException;
+import LuanFigueiraOliveira.ECommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
@@ -19,8 +24,9 @@ public class ProductService {
 
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Optional<Product> result = repository.findById(id);
-		Product product = result.get();
+		//Optional<Product> result = repository.findById(id);
+		//Product product = result.get();
+		Product product = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
 		return new ProductDTO(product);
 	}
 
@@ -45,15 +51,26 @@ public class ProductService {
 
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product product = repository.getReferenceById(id);
-		this.copyDtoToEntity(dto, product);
-		product = repository.save(product);
-		return new ProductDTO(product);
+		try {
+			Product product = repository.getReferenceById(id);
+			this.copyDtoToEntity(dto, product);
+			product = repository.save(product);
+			return new ProductDTO(product);
+		} catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 	}
 
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if(!repository.existsById(id)){
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+			repository.deleteById(id);
+		} catch(DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+		}
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product product) {
